Add files via upload
A lot has been changed and to be honest I don't quite understand what I did entirely. But what I do know is that my code can now read a file name and what version it is. Which I'm guessing is what is going to help when it comes to the different updates that are constantly made during production.